Add dropdown for mobile navigation
