+++ apps/builddao/widget/template/AppLayout.jsx
@@ -20,70 +20,6 @@ const ContentContainer = styled.div`
   width: 100%;
 `;
 
-const Navbar = styled.div`
-  width: 64px;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  padding: 48px;
-  width: 100%;
-
-  background-color: #0b0c14;
-`;
-
-const ButtonGroup = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 8px;
-`;
-
-const { NavLink } = props || {
-  NavLink: ({ to, children }) => (
-    <Link key={to} to={`/?page=${to}`}>
-      {children}
-    </Link>
-  ),
-};
-
-const AppHeader = ({ page, routes }) => (
-  <Navbar className="container-xl">
-    <img
-      style={{ width: 85, objectFit: "cover" }}
-      src="https://ipfs.near.social/ipfs/bafkreihbwho3qfvnu4yss3eh5jrx6uxhrlzdgtdjyzyjrpa6odro6wdxya"
-    />
-    <ButtonGroup>
-      {routes &&
-        (Object.keys(routes) || []).map((k) => {
-          const route = routes[k];
-          if (route.hide) {
-            return null;
-          }
-          return (
-            <NavLink to={k}>
-              <Button key={k} variant={page === k && "primary"}>
-                {route.init.icon && <i className={route.init.icon}></i>}
-                {route.init.name}
-              </Button>
-            </NavLink>
-          );
-        })}
-    </ButtonGroup>
-    <Widget
-      src="buildhub.near/widget/components.buttons.Connect"
-      props={{
-        connectedChildren: <div className="text-white">User Dropdown</div>,
-        showActivity: false,
-        className: "custom-button",
-        joinBtnChildren: "Join Now",
-        href: "/join",
-      }}
-    />
-  </Navbar>
-);
-
 const Footer = (props) => {
   return <></>;
 };
@@ -92,7 +28,11 @@ const Footer = (props) => {
 function AppLayout({ routes, page, children }) {
   return (
     <Container>
-      <AppHeader page={page} routes={routes} />
+      {/* <AppHeader page={page} routes={routes} /> */}
+      <Widget
+        src="buildhub.near/widget/components.navigation.header"
+        props={{ page, routes, ...props }}
+      />
       <ContentContainer key={page}>{children}</ContentContainer>
       <Footer page={page} />
     </Container>
